pyproject.toml
[project]
name = "gcloud-tts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.12",
    "google-cloud-texttospeech>=2.26.0",
    "uvicorn>=0.34.2",
]
